+++ src/utils/env.ts
@@ -3,6 +3,7 @@ import { bool, cleanEnv, num, str } from "envalid";
 const env = cleanEnv(process.env, {
   BOT_TOKEN: str(),
   CLIENT_ID: str(),
+  CLIENT_SECRET: str(),
   SPOTIFY_CLIENT_ID: str(),
   SPOTIFY_CLIENT_SECRET: str(),
   LAVALINK_NAME: str(),
@@ -12,6 +13,7 @@ const env = cleanEnv(process.env, {
   LAVALINK_SECURED: str(),
   DATABASE_URL: str(),
   DISABLE_UPDATE_COMMANDS: bool({ default: false }),
+  BASE_URL: str(),
 });
 
 export default env;